Refactored GuaranteePanel into GUMBOInterface
