+++ src/OSATE AADL AGREE Extension/src/org/osate/pluginsample/ui/GUMBOInterface.java
@@ -295,6 +295,145 @@ public class GUMBOInterface extends JFrame {
     	}
     }
     
+    private class GuaranteePanel extends JPanel {
+    	public GuaranteePanel() {
+    		setLayout(new BorderLayout());
+    		
+//    		List Panel
+    		JPanel listPanel = new JPanel();
+    		listPanel.setLayout(new FlowLayout());
+    		
+            DefaultListModel<String> listModel = new DefaultListModel<>();
+            for (int i = 0; i < 4; i++) {
+            	listModel.addElement(
+            			String.format("guarntee \"Example guarantee\" : (%s %s %d) %s %s", AGREEComponentFactory.getMockAssumptionParameter(), AGREEComponentFactory.getMockAssumptionComparator(), AGREEComponentFactory.getMockComparisonValue(), AGREEComponentFactory.getMockGuaranteeComparator(), AGREEComponentFactory.getMockGuaranteeParameter())
+    					);
+    		}
+            JList<String> guaranteeList = new JList<>(listModel); 
+            JScrollPane guaranteeListScrollPane = new JScrollPane(new JList<>(listModel));
+            listPanel.add(guaranteeListScrollPane);
+            
+            JButton removeGuaranteeButton = new JButton("-");
+            listPanel.add(removeGuaranteeButton);
+            
+    		add(listPanel, BorderLayout.PAGE_START);
+            
+//    		Inputs Panel
+            JPanel inputsPanel = new JPanel();
+            inputsPanel.setLayout(new BoxLayout(inputsPanel, BoxLayout.PAGE_AXIS));
+            
+            JTextField agreeDescriptionTextField = new JTextField();
+            inputsPanel.add(agreeDescriptionTextField);
+            
+            JComboBox<String> conditionalOperandList = new JComboBox<>(AGREEComponentFactory.getAllMockAssumptionParameters());
+            inputsPanel.add(conditionalOperandList);
+            
+//          1st Comparator Panel
+            JPanel assumptionComparatorPanel = new JPanel();
+            
+            JComboBox<String> assumptionComparatorList = new JComboBox<>(AGREEComponentFactory.getAllAssumptionComparators());
+            assumptionComparatorList.setMaximumSize( assumptionComparatorList.getPreferredSize());
+            assumptionComparatorPanel.add(assumptionComparatorList);
+            
+            inputsPanel.add(assumptionComparatorPanel);
+            
+            JTextField assumptionValueTextField = new JTextField();
+            inputsPanel.add(assumptionValueTextField);
+            
+//          2nd Comparator Panel
+            JPanel guaranteeComparatorPanel = new JPanel();
+            
+            JComboBox<String> guaranteeComparatorList = new JComboBox<>(AGREEComponentFactory.getAllGuaranteeComparators());
+            guaranteeComparatorList.setMaximumSize( guaranteeComparatorList.getPreferredSize());
+            guaranteeComparatorPanel.add(guaranteeComparatorList);
+            
+            inputsPanel.add(guaranteeComparatorPanel);
+            
+//          + Button Panel
+            JPanel addButtonPanel = new JPanel();
+            
+            JComboBox<String> guaranteeOperandList = new JComboBox<>(AGREEComponentFactory.getAllMockGuaranteeParameters());
+            addButtonPanel.add(guaranteeOperandList);
+            
+            JButton addGuaranteeButton = new JButton("+");
+            addButtonPanel.add(addGuaranteeButton);
+            
+            inputsPanel.add(addButtonPanel);
+            
+            add(inputsPanel);
+    	}
+    	
+    	private void updateListPane() {
+//    		if(assumptionListScrollPane != null) {
+//    			listPanel.remove(assumptionListScrollPane);
+//    		}
+//    		if(removeAssumptionButton != null) {
+//    			listPanel.remove(removeAssumptionButton);
+//    		}
+//    		
+//    		DefaultListModel<String> listModel = new DefaultListModel<>();
+//            for (String statement : assumptions) {
+//            	listModel.addElement(statement);
+//    		}
+//            
+//            assumptionList = new JList<>(listModel);
+//            assumptionListScrollPane = new JScrollPane(assumptionList);
+//            
+//            listPanel.add(assumptionListScrollPane);
+//            listPanel.add(removeAssumptionButton);
+            
+            revalidate();
+    		repaint();
+    	}
+    	
+    	private class RemoveGuaranteeAction extends AbstractAction {
+            public RemoveGuaranteeAction(String name) {
+                super(name);
+                int mnemonic = (int) name.charAt(0);
+                putValue(MNEMONIC_KEY, mnemonic);
+            }
+
+            @Override
+            public void actionPerformed(ActionEvent e) {
+//            	int currentSelection = assumptionList.getSelectedIndex();
+//            	
+//            	if(assumptionList.getModel().getSize() > 0
+//            		&& currentSelection > -1) {
+//	            	assumptions.remove(currentSelection);
+//	            	updateListPane();
+//            	}
+            }
+        }
+    	
+    	private class AddGuaranteeAction extends AbstractAction {
+    		public AddGuaranteeAction(String name) {
+    			super(name);
+                int mnemonic = (int) name.charAt(0);
+                putValue(MNEMONIC_KEY, mnemonic);
+    		}
+    		
+    		@Override
+            public void actionPerformed(ActionEvent e) {
+//    			String descValue = agreeDescriptionTextField.getText();
+//    			String parameterValue = assumptionOperandList.getSelectedItem().toString();
+//    			String comparatorValue = assumptionComparatorList.getSelectedItem().toString();
+//    			String assumptionValue = assumptionValueTextField.getText();
+//    			if(!descValue.isEmpty()
+//    				&& !parameterValue.isEmpty()
+//    				&& !comparatorValue.isEmpty()
+//					&& !assumptionValue.isEmpty()) {
+//	            	assumptions.add(String.format("assume \"%s\" : (%s %s %s)", descValue, parameterValue, comparatorValue, assumptionValue));
+//	            	updateListPane();
+//	            	
+//	            	agreeDescriptionTextField.setText("");
+//	            	assumptionValueTextField.setText("0");
+//    			} else {
+//    				
+//    			}
+            }
+    	}
+    }
+    
 	public static void main(String[] args) {
         new GUMBOInterface().setVisible(true);
     }
